Allowed DateTime object as input for UUID::mint(1) and fixed bugs
Instances implementing DateTimeInterface can now be used as input for
UUID::mint(1), but this is not recommended as microseconds are not
provided (even though documentation suggests they should be).

Fixed an unpacking bug when reusing clock sequence as well (oops), and
now throw a proper exception when a string time is not consistent with
the output of microtime().